backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.12
psycopg[binary]
